requirements.txt
numpy~=1.17.2
opencv-python~=4.1.1.26
torch~=1.5.0+cu92
torchvision~=0.6.0+cu92
matplotlib~=3.1.1
scipy~=1.4.1
rospy~=1.14.5
Pillow~=5.1.0